+++ master/buildbot/process/properties.py
@@ -13,10 +13,12 @@
 #
 # Copyright Buildbot Team Members
 
+import collections
 import re
 import weakref
 from buildbot import util
 from buildbot.interfaces import IRenderable, IProperties
+from twisted.internet import defer
 from twisted.python.components import registerAdapter
 from zope.interface import implements
 
@@ -134,7 +136,7 @@ class Properties(util.ComparableMixin):
 
     def render(self, value):
         renderable = IRenderable(value)
-        return renderable.getRenderingFor(self)
+        return defer.maybeDeferred(renderable.getRenderingFor, self)
 
 
 class PropertiesMixin:
@@ -288,102 +290,78 @@ class WithProperties(util.ComparableMixin):
             pmap.clear_temporary_values()
         return s
 
-class InterpolateMap(object): 
-    colon_minus_re = re.compile(r"(.*):-(.*)") 
-    colon_tilde_re = re.compile(r"(.*):~(.*)") 
-    colon_plus_re = re.compile(r"(.*):\+(.*)") 
-    def __init__(self, properties, kwargs): 
-        self.properties = properties 
-        self.kwargs = kwargs 
- 
-    def __getitem__(self, item): 
-        properties = self.properties 
 
-        key, arg = item.split(":", 1) 
 
-        def prop_fn(arg): 
-            try: 
-                prop, repl = arg.split(":", 1) 
-            except ValueError: 
-                prop, repl = arg, None 
-            return properties, prop, repl 
- 
-        def src_fn(arg): 
-            ## TODO: Handle changes 
-            try:
-                codebase, attr, repl = arg.split(":", 2)
-            except ValueError:
-                codebase, attr = arg.split(":",1)
-                repl = None
-            ss = self.properties.getBuild().getSourceStamp(codebase)
-            if ss:
-                return ss.asDict(), attr, repl
+_notHasKey = object() ## Marker object for _Lookup(..., hasKey=...) default
+class _Lookup(util.ComparableMixin):
+    implements(IRenderable)
+
+    def __init__(self, value, index, default=None,
+            defaultWhenFalse=True, hasKey=_notHasKey,
+            elideNoneAs=None):
+        self.value = value
+        self.index = index
+        self.default = default
+        self.defaultWhenFalse = defaultWhenFalse
+        self.hasKey = hasKey
+        self.elideNoneAs = elideNoneAs
+
+    @defer.inlineCallbacks
+    def getRenderingFor(self, build):
+        value = build.render(self.value)
+        index = build.render(self.index)
+        value, index = yield defer.gatherResults([value, index])
+        if not value.has_key(index):
+           rv = yield build.render(self.default)
+        else:
+            if self.defaultWhenFalse:
+                rv = yield build.render(value[index])
+                if not rv:
+                    rv = yield build.render(self.default)
+                elif self.hasKey is not _notHasKey:
+                    rv = yield build.render(self.hasKey)
+            elif self.hasKey is not _notHasKey:
+                rv = yield build.render(self.hasKey)
             else:
-                return {}, attr, repl
+                rv = yield build.render(value[index])
+        if rv is None:
+            rv = yield build.render(self.elideNoneAs)
+        defer.returnValue(rv)
+
+
+def _getInterpolationList(fmtstring):
+    # TODO: Verify that no positial substitutions are requested
+    dd = collections.defaultdict(str)
+    fmtstring % dd
+    return dd.keys()
+
+
+class _PropertyDict(object):
+    implements(IRenderable)
+    def getRenderingFor(self, build):
+        return build.getProperties()
+_thePropertyDict = _PropertyDict()
+
+class _SourceStampDict(object):
+    implements(IRenderable)
+    def __init__(self, codebase):
+        self.codebase = codebase
+    def getRenderingFor(self, build):
+        ss = build.getBuild().getSourceStamp(self.codebase)
+        if ss:
+            return ss.asDict()
+        else:
+            return {}
+
+
+class _Lazy(object):
+    implements(IRenderable)
+    def __init__(self, value):
+        self.value = value
+    def getRenderingFor(self, build):
+        return self.value
  
-        def kw_fn(arg): 
-            try: 
-                kw, repl = arg.split(":", 1) 
-            except ValueError: 
-                kw, repl = arg, None 
-            return self.kwargs, kw, repl 
- 
-        for (keys, fn) in [ 
-            (("p", "prop"), prop_fn), 
-            (("s", "src"), src_fn), 
-            (("kw", "kw"), kw_fn), 
-            ]: 
-            if key in keys: 
-                d, kw, repl = fn(arg) 
-                break;
- 
-        def colon_minus(d, kw, repl): 
-            # %(prop:-repl)s 
-            # if prop exists, use it; otherwise, use repl 
-            if d.has_key(kw): 
-                return d[kw] 
-            else: 
-                return repl % self 
- 
-        def colon_tilde(d, kw, repl): 
-            # %(prop:~repl)s 
-            # if prop exists and is true (nonempty), use it; otherwise, use repl 
-            if d.has_key(kw) and d[kw]: 
-                return d[kw] 
-            else: 
-                return repl % self 
- 
-        def colon_plus(d, kw, repl): 
-            # %(prop:+repl)s 
-            # if prop exists, use repl; otherwise, an empty string 
-            if d.has_key(kw): 
-                return repl % self 
-            else: 
-                return '' 
- 
-        if repl is not None: 
-            good_colon = False 
-            for char, fn in [ 
-                ( "-", colon_minus ), 
-                ( "~", colon_tilde ), 
-                ( "+", colon_plus ), 
-                ]: 
-                if repl[0] == char: 
-                    rv = fn(d, kw, repl[1:]) 
-                    good_colon = True 
-                    break 
-            if not good_colon: 
-                raise ValueError 
-        else: 
-            if d.has_key(kw): 
-                rv = d[kw] 
-            else: 
-                rv = None 
- 
-        # translate 'None' to an empty string 
-        if rv is None: rv = '' 
-        return rv 
- 
+
 class Interpolate(util.ComparableMixin): 
     """ 
     This is a marker class, used fairly widely to indicate that we 
@@ -394,20 +372,111 @@ class Interpolate(util.ComparableMixin):
     compare_attrs = ('fmtstring', 'args', 'kwargs') 
  
     def __init__(self, fmtstring, *args, **kwargs): 
+        global config
+        from buildbot import config
         self.fmtstring = fmtstring 
-        self.args = args 
+        self.args = args
         self.kwargs = kwargs 
-        if self.args and self.kwargs: 
-            raise ValueError('Interpolate takes either positional or keyword substitutions, not both.') 
+        if not self.args:
+            self.interpolations = {}
+            self._parse(fmtstring)
+            keys = _getInterpolationList(fmtstring)
+        if self.args and self.kwargs:
+            raise ValueError('Interpolate takes either positional or keyword substitutions, not both.')
+
+    @staticmethod
+    def _parse_prop(arg):
+        try:
+            prop, repl = arg.split(":", 1)
+        except ValueError:
+            prop, repl = arg, None
+        return _thePropertyDict, prop, repl
+
+    @staticmethod
+    def _parse_src(arg):
+        ## TODO: Handle changes
+        try:
+            codebase, attr, repl = arg.split(":", 2)
+        except ValueError:
+            try:
+                codebase, attr = arg.split(":",1)
+                repl = None
+            except ValueError:
+                config.error("Must specify both codebase and attribute for src Interpolation '%s'" % arg)
+                codebase = attr = repl = None
+        return _SourceStampDict(codebase), attr, repl
+
+    def _parse_kw(self, arg):
+        try:
+            kw, repl = arg.split(":", 1)
+        except ValueError:
+            kw, repl = arg, None
+        return _Lazy(self.kwargs), kw, repl
+
+    def _parseSubstitution(self, fmt):
+        try:
+            key, arg = fmt.split(":", 1)
+        except ValueError:
+            config.error("invalid Interpolate substitution without selector '%s'" % fmt)
+            return
+
+        fn = getattr(self, "_parse_" + key, None)
+        if not fn:
+            config.error("invalid Interpolate selector '%s'" % key)
+            return None
+        else:
+            return fn(arg)
+
+    def _parseColon_minus(self, d, kw, repl):
+        return _Lookup(d, kw,
+               default=Interpolate(repl, **self.kwargs),
+               defaultWhenFalse=False,
+               elideNoneAs='')
+
+    def _parseColon_tilde(self, d, kw, repl):
+        return _Lookup(d, kw,
+               default=Interpolate(repl, **self.kwargs),
+               defaultWhenFalse=True,
+               elideNoneAs='')
+
+    def _parseColon_plus(self, d, kw, repl):
+        return _Lookup(d, kw,
+               hasKey=Interpolate(repl, **self.kwargs),
+               default='',
+               defaultWhenFalse=False,
+               elideNoneAs='')
+
+    def _parse(self, fmtstring):
+        keys = _getInterpolationList(fmtstring)
+        for key in keys:
+            if not self.interpolations.has_key(key):
+                d, kw, repl = self._parseSubstitution(key)
+                if repl is None:
+                    repl = '-'
+                for char, fn in [
+                    ( "-", self._parseColon_minus ),
+                    ( "~", self._parseColon_tilde ),
+                    ( "+", self._parseColon_plus ),
+                    ]:
+                    if repl[0] == char:
+                        self.interpolations[key] = fn(d, kw, repl[1:])
+                if not self.interpolations.has_key(key):
+                    config.error("invalid Interpolate default type '%s'" % repl[0])
 
     def getRenderingFor(self, props): 
         props = props.getProperties() 
         if self.args:
-            args = props.render(self.args) 
-            return self.fmtstring % tuple(args) 
+            d = props.render(self.args)
+            @d.addCallback
+            def format(args):
+                return self.fmtstring % tuple(args)
+            return d
         else:
-            kwargs = props.render(self.kwargs) 
-            return self.fmtstring % InterpolateMap(props, kwargs) 
+            d = props.render(self.interpolations)
+            @d.addCallback
+            def format(res):
+                return self.fmtstring % res
+            return d
 
 class Property(util.ComparableMixin):
     """
@@ -432,14 +501,19 @@ class Property(util.ComparableMixin):
 
     def getRenderingFor(self, props):
         if self.defaultWhenFalse:
-            rv = props.getProperty(self.key)
-            if rv:
-                return rv
+            d = props.render(props.getProperty(self.key))
+            @d.addCallback
+            def checkDefault(rv):
+                if rv:
+                    return rv
+                else:
+                    return props.render(self.default)
+            return d
         else:
             if props.hasProperty(self.key):
-                return props.getProperty(self.key)
-
-        return props.render(self.default)
+                return props.render(props.getProperty(self.key))
+            else:
+                return props.render(self.default)
 
 class _DefaultRenderer(object):
     """
@@ -472,7 +546,7 @@ class _ListRenderer(object):
         self.value = value
 
     def getRenderingFor(self, build):
-        return [ build.render(e) for e in self.value ]
+        return defer.gatherResults([ build.render(e) for e in self.value ])
 
 registerAdapter(_ListRenderer, list, IRenderable)
 
@@ -488,7 +562,9 @@ class _TupleRenderer(object):
         self.value = value
 
     def getRenderingFor(self, build):
-        return tuple([ build.render(e) for e in self.value ])
+        d = defer.gatherResults([ build.render(e) for e in self.value ])
+        d.addCallback(tuple)
+        return d
 
 registerAdapter(_TupleRenderer, tuple, IRenderable)
 
@@ -501,10 +577,11 @@ class _DictRenderer(object):
     implements(IRenderable)
 
     def __init__(self, value):
-        self.value = value
+        self.value = _ListRenderer([ _TupleRenderer((k,v)) for k,v in value.iteritems() ])
 
     def getRenderingFor(self, build):
-        return dict([ (build.render(k), build.render(v)) for k,v in self.value.iteritems() ])
-
+        d = self.value.getRenderingFor(build)
+        d.addCallback(dict)
+        return d
 
 registerAdapter(_DictRenderer, dict, IRenderable)