+++ master/buildbot/test/unit/test_steps_source_base_Source.py
@@ -17,7 +17,7 @@ import mock
 from twisted.trial import unittest
 
 from buildbot.steps.source import Source
-from buildbot.test.util import sourcesteps
+from buildbot.test.util import steps, sourcesteps
 
 class TestSource(sourcesteps.SourceStepMixin, unittest.TestCase):
 
@@ -65,3 +65,26 @@ class TestSource(sourcesteps.SourceStepMixin, unittest.TestCase):
         step.startStep(mock.Mock())
 
         self.assertEqual(step.startVC.call_args, (('branch', None, None), {}))
+
+
+class TestSourceDescription(steps.BuildStepMixin, unittest.TestCase):
+
+    def setUp(self):
+        return self.setUpBuildStep()
+
+    def tearDown(self):
+        return self.tearDownBuildStep()
+
+    def test_constructor_args_strings(self):
+        step = Source(workdir='build',
+                      description='svn update (running)',
+                      descriptionDone='svn update')
+        self.assertEqual(step.description, ['svn update (running)'])
+        self.assertEqual(step.descriptionDone, ['svn update'])
+
+    def test_constructor_args_lists(self):
+        step = Source(workdir='build',
+                      description=['svn', 'update', '(running)'],
+                      descriptionDone=['svn', 'update'])
+        self.assertEqual(step.description, ['svn', 'update', '(running)'])
+        self.assertEqual(step.descriptionDone, ['svn', 'update'])